Deadwood Trees
Added Deadwood, a new type of tree that replaces the Dark Oak and Swamp trees that used to be in the Spooky Dimension.
With textures by MoonlitLamps, these trees come in 2 sizes and look fantastic!
They have everything that vanilla woods have except boats, because boats are complicated.
Other changes include:
- Added Eerie Grass, replaces vanilla grass in the spooky dimension
- Infested Tainted Stone now spawns underground in the spooky dimension
- Added more things to the non-US Englishes that were missing
- Fixed the log complaining about Witch Water Bubble Columns not having a blockstate json
- Removed the item form of WItch Water Bubble Columns because why did that exist anyway
- Ran another cleanup