1.15 update part 1: fixed compile errors in blocks.
Notable changes:
*onUse now returns an ActionResult instead of a boolean.
*Signs now take a WoodType in the constructor. Thankfully, WoodType is
not an enum. I've added an inner class to TheHallow that extends
WoodType to access its protected constructor.
*Block render layers are now a hashmap instead of a method override.
This means that we can do away with TranslucentGlassBlock and
TranslucentGlassPane (though HallowedPaneBlock is necessary to access
PaneBlock's protected constructor.)
*isOpaque is now a final method and can be set via the Block.Settings.

The rest is mostly renames and minor refactors from Mojang.